[#10] Improving folder structure
