+++ index.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const cors = require('cors');
 const db = require('./dbconfig/dbConfig');
-const router = require('./routes');
+const configRoutes = require('./src/routes/router');
 
 var app = express()
 app.use(cors())
@@ -9,6 +9,7 @@ app.use(cors())
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 
-app.use('/', router)
+require('./src/routes/router')(app);
+configRoutes(app);
 
 app.listen(3001)
